pyproject.toml
[tool.poetry]
name = "git-kpis"
version = "0.1.0"
description = "A tool for generating lead time based on GitHub commits that match a pattern and release tags"
authors = ["Adrian Chung <adrian.c.chung@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.21"
pendulum = "^2.0"
numpy = "^1.16"
pygithub = "^1.43"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pre-commit = "^1.15"
pytest-cov = "^2.6"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[tool.poetry.scripts]
metrics = 'git_kpis.git_kpis:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
